# server/src/docs/swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: Scan APIs
  description: Scan APIs
  license:
    name: Github
    url: 'https://github.com/rupayaproject/scan'
...
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/tokens:
    get:
      tags:
        - Token
      summary: Get list tokens
      parameters:
        - name: type
          in: query
          required: false
          description: trc20|trc21|trc721, default = trc20
          type: string
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
  /api/tokens/{hash}:
    get:
      tags:
        - Tokens
      summary: Get token detail
      parameters:
        - name: hash
          in: path
          required: true
          description: token address
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
  /api/tokens/holding/{tokenType}/{holder}:
    get:
      tags:
        - Tokens
      summary: Get list token holding of a holder
      parameters:
        - name: tokenType
          in: path
          required: true
          description: trc20|trc21|trc721
          type: string
        - name: holder
          in: path
          required: true
          description: holder address
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
...
  /api/token-txs/{tokenType}:
    get:
      tags:
        - Token
      summary: Get list token transactions.
      parameters:
        - name: tokenType
          in: path
          required: true
          description: trc20|trc21|trc721
          type: string
        - name: holder
          in: query
          required: false
          description: holder address
          type: string
        - name: token
          in: query
          required: false
          description: token address
          type: string
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
  /api/token-txs/{tokenType}/{txHash}:
    get:
      tags:
        - Token
      summary: Get list token transactions by transaction hash.
      parameters:
        - name: tokenType
          in: path
          required: true
          description: trc20|trc21|trc721
          type: string
        - name: txHash
          in: path
          required: false
          description: token address
          type: string
        - name: holder
          in: query
          required: false
          description: holder address
          type: string
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
...